Fixed import of children of children, cleaned up and refactored code, introduced hard blacklist to ignore entities from site.

+++ populate_database_sql.py
@@ -1,22 +1,18 @@
-'''
-Ankur Jai Sood
-29/5/2015
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
 
-Populate_database.py
-Function:
-1. Performs search of CHEBI database for DNA bases
-2. Returns chebiId and chebiAsciiName of bases
-3. Searches CHEBI database for all entities of which
-   the DNA bases are functional parents
-4. Using above results, populates DNA post-transciptional modification table
-'''
+from __future__ import with_statement, division, print_function
 
 # Using Suds web services client for soap
+import csv
 import os
 import sqlite3
+import sys
 from suds.client import Client
-from sys import maxint
 from Bio import Entrez
+from sys import maxint
+
+import dnamod_utils
 
 # Search Variables made up of CHEBI object attributes
 SYNONYM_SEARCH_STR = 'Synonyms'
@@ -31,18 +27,13 @@ ONTOLOGY_HAS_ROLE = "has role"
 RESET_TABLES = True
 
 # Program Constants
+BLACK_LIST = []
 DNA_BASES = ['cytosine', 'thymine', 'adenine', 'guanine', 'uracil']
 FILE_PATH = os.path.dirname(os.path.abspath(__file__))
-print FILE_PATH
+print("Operating from: {}".format(FILE_PATH))
 
-# Url to CHEBI wsdl schema
 url = 'https://www.ebi.ac.uk/webservices/chebi/2.0/webservice?wsdl'
-
-# Create a client object from above url
 client = Client(url)
-
-# This line prints out all methods and data types available from webservice
-# print client
 
 
 def search_exact_match(searchKey, client):
@@ -52,14 +43,9 @@ def search_exact_match(searchKey, client):
     maximumResults = maxint
     starsCategory = 'THREE ONLY'
 
-    # Query CHEBI using above key and store in list results
-    # getLiteEntity(xs:string search, SearchCategory searchCategory,
-    #               xs:int maximumResults, StarsCategory stars)
-
     # Save results from query into list
     results = client.service.getLiteEntity(searchKey, searchCategory,
                                            maximumResults, starsCategory)
-
     if not results:
         result = 'Invalid Input'
         return result
@@ -72,61 +58,80 @@ def search_exact_match(searchKey, client):
     result = 'DNE'
 
     # Iterate through all of the results and find the actual base
-    for elements in range(len(results)):
-        if results[elements].chebiAsciiName == searchKey:
-            result = results[elements]
+    for entities in results:
+        if entities.chebiAsciiName == searchKey:
+            result = entities
 
     return result
 
 
 def search_for_bases(client):
-    # Initialize DNA bases
-    dnaBases = DNA_BASES
-
-    # Code for debugging
-    # print dnaBases[0]
-    # print len(dnaBases)
-    # print range(len(dnaBases))
-
     # Initialize empy list
     result = []
 
     # Search CHEBI for bases and return lite entries
-    for elements in range(len(dnaBases)):
+    for base in DNA_BASES:
         # print elements # Output for debugging
-        result.append(search_exact_match(dnaBases[elements], client))
+        result.append(search_exact_match(base, client))
 
     return result
 
 
 def filter_stars(content, stars, client):
     result = []
-    for elements in range(len(content)):
+    for entity in content:
         # print content[elements] # For Debugging
         # content[elements] = search_exact_match(content[elements].chebiName,
         #                                        client)
-        temphold = content[elements]
+        temphold = entity
         # print temphold # For debugging
-        content[elements] = get_complete_entity(content[elements].chebiId,
-                                                client)
+        #time.sleep(1)
+        entity = get_complete_entity(entity.chebiId, client)
         # print content[elements] # For Debugging
-        if content[elements] == 'Invalid Input' or content[elements] == 'DNE':
+        if entity == 'Invalid Input' or entity == 'DNE':
             continue
-        elif (content[elements].entityStar == stars and
-              temphold.type == 'has functional parent'):
-            result.append(content[elements])
+        elif (entity.entityStar == stars and
+              (temphold.type == 'has functional parent' 
+               or temphold.type == 'is a')):
+            result.append(entity)
     return result
 
 
 def get_children(bases, client):
     modDictionary = {}
-    for elements in range(len(bases)):
-        result = client.service.getOntologyChildren(bases[elements].chebiId)
+    for base in bases:
+        result = client.service.getOntologyChildren(base.chebiId)
+        print("----- BASE: {}".format(base.chebiAsciiName))
         result = result.ListElement
         result = filter_stars(result, 3, client)
-        # print bases[elements] # Debugging output
-        modDictionary[bases[elements].chebiAsciiName] = result
+        additionalChildren = get_further_children(result, client)
+
+        for child in additionalChildren:
+            result.append(child)
+
+        result = set(result)
+        result = list(result)
+
+        modDictionary[base.chebiAsciiName] = result
+
     return modDictionary
+
+
+def get_further_children(entities, client):
+    additionalChildren = []
+    for entity in entities:
+        if entity.chebiAsciiName not in BLACK_LIST:
+            print("---------- CHILD of BASE: {}".format(entity.chebiAsciiName))
+            result = client.service.getOntologyChildren(entity.chebiId)
+            if result:
+                result = result.ListElement
+                result = filter_stars(result, 3, client)
+                recursivestep = get_further_children(result, client)
+                result = result + recursivestep
+                result = set(result)
+                result = list(result)
+                additionalChildren.extend(result)
+    return additionalChildren
 
 
 def get_complete_entity(CHEBIid, client):
@@ -136,8 +141,8 @@ def get_complete_entity(CHEBIid, client):
 
 def get_complete_bases(bases, client):
     result = []
-    for elements in range(len(bases)):
-        result.append(get_complete_entity(bases[elements].chebiId, client))
+    for base in bases:
+        result.append(get_complete_entity(base.chebiId, client))
     return result
 
 
@@ -145,7 +150,9 @@ def create_base_table(bases):
     conn = sqlite3.connect('DNA_mod_database.db')
     c = conn.cursor()
 
-    c.execute('''DROP TABLE IF EXISTS base''')
+    if RESET_TABLES is True:
+        c.execute('''DROP TABLE IF EXISTS base''')
+
     c.execute('''CREATE TABLE IF NOT EXISTS base
                 (baseid TEXT PRIMARY KEY NOT NULL,
                  commonname text,
@@ -153,7 +160,9 @@ def create_base_table(bases):
     conn.commit()
 
     for base in bases:
-        c.execute("INSERT INTO base VALUES(?,?,?)",
+        c.execute("UPDATE base SET commonname=?, basedefinition=? WHERE commonname=?",(base.chebiAsciiName, base.chebiAsciiName, base.chebiAsciiName[0]))
+        conn.commit()
+        c.execute("INSERT OR IGNORE INTO base VALUES(?,?,?)",
                   (base.chebiAsciiName[0],
                    base.chebiAsciiName, base.definition))
     conn.commit()
@@ -161,29 +170,35 @@ def create_base_table(bases):
 
 
 def concatenate_list(child, attribute):
-    return ([synonym_data.data
-             for synonym_data in getattr(child, attribute)]
-            if attribute in dir(child) else [])
+    if attribute in dir(child):
+        return [synonym_data.data for synonym_data in getattr(child, attribute)]
+    else:
+        return []
 
 
 def get_entity(child, attribute):
-    return ([getattr(child, attribute)]
-            if attribute in dir(child) else [])
-
+    if attribute in dir(child):
+        return ([getattr(child, attribute)])
+    else:
+        return []
 
 def get_roles(child, attribute, selector):
-    return ([ontologyObject
-             for ontologyObject in getattr(child, attribute)
-            if ontologyObject.type == ONTOLOGY_HAS_ROLE])
+    return [ontologyitem for ontologyitem in
+            getattr(child, attribute) if ontologyitem.type == selector] 
 
 
 def get_full_citation(PMID):
+    print("Adding Citation: ",PMID)
     result = []
     # isbook = False # Unsed at the moment
     isarticle = False
     Entrez.email = "jai.sood@hotmail.com"
     handle = Entrez.efetch("pubmed", id=PMID, retmode="xml")
     records = Entrez.parse(handle)
+
+    articleTitle = []
+    publicationDate = []
+    authors = []
 
     for record in records:
         if 'MedlineCitation' in record.keys():
@@ -219,33 +234,31 @@ def get_full_citation(PMID):
 
     handle.close()
 
-    if articleTitle is not None and articleTitle != []:
-        result.append(articleTitle)
+    if articleTitle:
+        result.append(articleTitle.encode('utf-8'))
     else:
         result.append('Title Not Found')
 
     if isarticle:
-        if publicationDate is not None and publicationDate != []:
+        if publicationDate:
             date = (publicationDate[0]['Month'] + '-' +
                     publicationDate[0]['Day'] + '-' +
                     publicationDate[0]['Year'])
-            result.append(str(date))
+            result.append(date)
         else:
             result.append('Publication Date Not Found')
     else:
-        if publicationDate is not None and publicationDate != []:
+        if publicationDate:
             date = (publicationDate['Month'] + '-' + publicationDate['Day']
                     + '-' + publicationDate['Year'])
             result.append(date)
         else:
             result.append('Publication Date Not Found')
-
-    if authors is not None and authors != []:
-        name = (authors[0]['LastName'] + ', ' + authors[0]['Initials'] +
-                '. et al.')
-        result.append(name)
+    if authors:
+        result.append("{0}, {1}, et al.".format(authors[0]['LastName'].encode("utf-8"),
+                                              authors[0]['Initials'].encode("utf-8")))
     else:
-        result.append('Author Not Found')
+        result.append('Authors Not Found')
 
     return result
 
@@ -350,8 +363,11 @@ def create_other_tables(children, bases):
                           (role_ids[role],))
                 data = c.fetchone()[0]
                 if data == 0:
-                    c.execute("INSERT INTO roles VALUES(?,?)",
+                    c.execute("UPDATE roles SET role=? WHERE roleid=?",(role_names[role], role_ids[role]))
+                    conn.commit()
+                    c.execute("INSERT OR IGNORE INTO roles VALUES(?,?)",
                               (role_ids[role], role_names[role]))
+                    conn.commit()
 
             for citation in citations:
                 c.execute("SELECT * FROM citations WHERE citationid = ?",
@@ -359,27 +375,33 @@ def create_other_tables(children, bases):
                 data = c.fetchone()
                 if data is None:
                     citationinfo = get_full_citation(citation)
-                    c.execute("INSERT INTO citations VALUES(?,?,?,?)",
-                              (citation, citationinfo[0], citationinfo[1],
-                               citationinfo[2]))
+                    citationinfo_uni = [info.decode('utf-8') for info in citationinfo]
 
-            c.execute("INSERT INTO baseprops VALUES(NULL,?,?,?)",
+                    c.execute("UPDATE citations SET title=?, pubdate=?, authors=? WHERE citationid=?",(citationinfo_uni[0], citationinfo_uni[1], citationinfo_uni[2], citation))
+                    conn.commit()
+                    c.execute("INSERT OR IGNORE INTO citations VALUES(?,?,?,?)",
+                              (citation, citationinfo_uni[0], citationinfo_uni[1],
+                               citationinfo_uni[2]))
+                    conn.commit()
+
+            c.execute("INSERT OR IGNORE INTO baseprops VALUES(NULL,?,?,?)",
                       (str(formula), str(charge), str(mass)))
-            c.execute("INSERT INTO names VALUES(NULL,?,?,?,?,?)",
+            c.execute("INSERT OR IGNORE INTO names VALUES(NULL,?,?,?,?,?)",
                       (child.chebiAsciiName, child.chebiId,
                        str(iupac), str(synonyms), str(smiles)))
-            c.execute("INSERT INTO covmod VALUES(NULL,?,?)",
+            c.execute("INSERT OR IGNORE INTO covmod VALUES(NULL,?,?)",
                       ('0', child.definition))
             rowid = c.lastrowid
 
             for role in range(len(roles)):
-                c.execute("INSERT INTO roles_lookup VALUES(?,?)",
+                c.execute("INSERT OR IGNORE INTO roles_lookup VALUES(?,?)",
                           (rowid, role_ids[role]))
+
             for citation in citations:
-                c.execute("INSERT INTO citation_lookup VALUES(?,?)",
+                c.execute("INSERT OR IGNORE INTO citation_lookup VALUES(?,?)",
                           (rowid, citation))
 
-            c.execute("INSERT INTO modbase VALUES(NULL,?,?,?,?,?,?,?)",
+            c.execute("INSERT OR IGNORE INTO modbase VALUES(NULL,?,?,?,?,?,?,?)",
                       ('0', base.chebiAsciiName[0], rowid,
                        rowid, rowid, rowid, rowid))
             conn.commit()
@@ -391,26 +413,13 @@ def populate_tables(bases, children, client):
     create_base_table(bases)
     create_other_tables(children, bases)
 
-
-# 1. Performs search of CHEBI database for DNA bases
-# 2. Returns chebiId and chebiAsciiName of bases
-print "1/4 Searching for bases..."
+BLACK_LIST = dnamod_utils.get_list('blacklist')
+print("1/4 Searching for bases...")
 bases = search_for_bases(client)
-# print bases # Debugging Output
-
-# 3. Searches CHEBI database for all entities of which
-#    the DNA bases are functional parents and are rated three or more stars
-print "2/4 Searching for children..."
+print("2/4 Searching for children...")
 children = get_children(bases, client)
 bases = get_complete_bases(bases, client)
-# print children.keys() # Debugging output
-# print bases[0] # Debugging output
-# print children[bases[0].chebiAsciiName][0]  # Debugging output
-# print children[bases[0].chebiAsciiName]  # Debugging output
-# print children.items()
-
-# 4. Using above results, populates DNA post-transciptional modification table
-print "3/4 Creating tables..."
+print("3/4 Creating tables...")
 populate_tables(bases, children, client)
-print "4/4 Finishing up..."
-print "Done!"
+print("4/4 Finishing up...")
+print("Done!")
